+++ R/manynet-utils.R
@@ -45,6 +45,10 @@ interpolate <- function(values, positions, n, fill = NA) {
   out
 }
 
+is.scalar <- function(x) {
+  is.atomic(x) && length(x) == 1L
+}
+
 # #' @export
 # `%||%` <- function(x, y) {
 #   if (is_null(x)) y else x